# openapi.yaml
openapi: 3.1.0
info:
  title: MOA API Contract
  description: Single Source of Truth for MOA Spring Boot 3 Backend & Flutter Frontend REST API
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "POST /auth/login 응답의 accessToken을 'Authorization: Bearer <token>'으로 전달한다."
  schemas:
    ApiError:
      type: object
      properties:
        code:
          type: string
          example: INVALID_AUTH_TOKEN
        message:
          type: string
        details:
          type: array
          items:
            type: string
...
    Club:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        leaderName:
          type: string
        category:
          type: string
        memberCount:
          type: integer
        thumbnailUrl:
          type: string
          nullable: true
        joined:
          type: boolean
          description: 로그인한 사용자가 이 동아리에 가입했는지 여부.
        favorite:
          type: boolean
          description: 로그인한 사용자가 이 동아리를 즐겨찾기했는지 여부. joined가 false면 항상 false다.
        leader:
          type: boolean
          description: >-
            로그인한 사용자가 이 동아리의 관리자(동아리장)인지 여부. clubs.leader_id와
            로그인한 사용자 ID를 비교해서 서버가 판단한다 — leaderName(표시용
            문자열) 비교가 아니라 실제 FK 기준이라 동명이인이 있어도 정확하다.
            joined가 false면 항상 false다.
    SetClubFavoriteRequest:
      type: object
      required: [favorite]
      properties:
        favorite:
          type: boolean
    TransferClubLeaderRequest:
      type: object
      required: [newLeaderId]
      properties:
        newLeaderId:
          type: integer
          format: int64
          description: 관리자 권한을 넘겨받을 대상의 User ID. 반드시 이 동아리의 기존 가입 멤버여야 한다.
    ClubMemberSummary:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        name:
          type: string
        leader:
          type: boolean
          description: 이 멤버가 현재 동아리장인지 여부.
    ClubApplication:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        applicantName:
          type: string
        selfIntroduction:
          type: string
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        appliedAt:
          type: string
          format: date-time
    ClubDetail:
      allOf:
        - $ref: '#/components/schemas/Club'
        - type: object
          properties:
            description:
              type: string
              nullable: true
              description: 가입 전 소개(동아리 상세 화면)에만 노출되는 소개 글.
            applicationStatus:
              type: string
              enum: [PENDING, REJECTED]
              nullable: true
              description: >-
                joined가 true면 항상 null이다. joined가 false일 때만
                null(아직 신청 안 함)/PENDING(승인 대기)/REJECTED(거절됨,
                재신청 가능) 중 하나가 온다.
    ApplyClubRequest:
      type: object
      required: [selfIntroduction]
      properties:
        selfIntroduction:
          type: string
          minLength: 20
          maxLength: 1000
...
paths:
...
  /clubs/me:
    get:
      summary: 내가 가입한 동아리 목록
      operationId: getMyClubs
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Club'
                  error:
                    nullable: true
        '401':
          description: 인증되지 않음
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
  /clubs:
    get:
      summary: 전체 동아리 목록 (메인 페이지 홈 피드)
      description: >-
        로그인한 사용자 기준으로 각 동아리의 joined/favorite 여부를 함께 내려준다.
      operationId: getAllClubs
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Club'
                  error:
                    nullable: true
        '401':
          description: 인증되지 않음
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
    post:
      summary: 스터디(동아리) 등록
      description: >-
        메인 페이지 "스터디 등록" 버튼 → 등록 화면의 "작성완료" 제출. 사진(thumbnail)
        파일을 함께 받기 위해 JSON이 아니라 multipart/form-data로 받는다. 등록한
        사용자가 곧바로 그 스터디의 회장 겸 첫 멤버가 된다(joined=true로 응답).
        name이 비어있거나 100자를 넘으면 INVALID_CLUB_NAME(400), 이미 존재하는
        이름이면 DUPLICATE_CLUB_NAME(409)을 반환한다.
      operationId: createClub
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - name
                - description
              properties:
                name:
                  type: string
                  maxLength: 100
                description:
                  type: string
                thumbnail:
                  type: string
                  format: binary
      responses:
        '200':
          description: 등록 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/ClubDetail'
                  error:
                    nullable: true
        '400':
          description: 이름이 비어있거나 형식이 올바르지 않음 (INVALID_CLUB_NAME)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증되지 않음
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
        '409':
          description: 이미 존재하는 스터디 이름 (DUPLICATE_CLUB_NAME)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
  /clubs/{clubId}/favorite:
    patch:
      summary: 동아리 즐겨찾기 설정
      description: 가입한 동아리에 대해서만 즐겨찾기를 설정할 수 있다.
      operationId: setClubFavorite
      security:
        - bearerAuth: []
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetClubFavoriteRequest'
      responses:
        '200':
          description: 변경 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Club'
                  error:
                    nullable: true
        '401':
          description: 인증되지 않음
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
        '404':
          description: 존재하지 않는 동아리이거나(CLUB_NOT_FOUND), 가입하지 않은 동아리를 즐겨찾기하려 함(CLUB_MEMBERSHIP_NOT_FOUND)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
  /clubs/{clubId}:
    get:
      summary: 동아리 상세(가입 전 소개/지원) 조회
      operationId: getClubDetail
      security:
        - bearerAuth: []
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/ClubDetail'
                  error:
                    nullable: true
        '401':
          description: 인증되지 않음
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
        '404':
          description: 존재하지 않는 동아리(CLUB_NOT_FOUND)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
  /clubs/{clubId}/apply:
    post:
      summary: 동아리 가입 신청 ('지원 하기')
      description: >-
        가입하지 않은 동아리에 자기소개와 함께 가입을 신청한다. 신청은 PENDING
        상태로 생성되며, 동아리장 승인/거절 기능은 아직 없다(이번 범위 밖).
        REJECTED 상태의 기존 신청서가 있으면 재신청으로 처리되어 같은 신청서가
        새 자기소개로 갱신되고 다시 PENDING이 된다.
      operationId: applyToClub
      security:
        - bearerAuth: []
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplyClubRequest'
      responses:
        '200':
          description: 신청 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/ClubDetail'
                  error:
                    nullable: true
        '400':
          description: 요청 값이 올바르지 않음 (자기소개 누락/길이 미달)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증되지 않음
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
        '404':
          description: 존재하지 않는 동아리(CLUB_NOT_FOUND)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
        '409':
          description: 이미 가입함(CLUB_ALREADY_JOINED) 또는 이미 승인 대기 중인 신청서가 있음(CLUB_APPLICATION_ALREADY_PENDING)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
  /clubs/{clubId}/members:
    get:
      summary: 동아리 멤버 목록
      description: >-
        관리자 권한 넘기기 화면(멤버 선택)에서 쓴다. 가입한 사용자만 조회할 수
        있다 — 가입하지 않은 동아리를 조회하면 CLUB_MEMBERSHIP_NOT_FOUND(404).
      operationId: getClubMembers
      security:
        - bearerAuth: []
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ClubMemberSummary'
                  error:
                    nullable: true
        '401':
          description: 인증되지 않음
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
        '404':
          description: 존재하지 않는 동아리(CLUB_NOT_FOUND) 또는 가입하지 않음(CLUB_MEMBERSHIP_NOT_FOUND)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
  /clubs/{clubId}/leader:
    patch:
      summary: 관리자(동아리장) 권한 넘기기
      description: >-
        현재 동아리장만 호출할 수 있다. 대상(newLeaderId)은 반드시 이 동아리의
        기존 가입 멤버여야 한다. 성공하면 clubs.leader_id와 leaderName이 모두
        새 동아리장 기준으로 바뀌고, 응답의 leader는 호출한 사용자(이제는 더
        이상 동아리장이 아님) 기준으로 false가 된다.
      operationId: transferClubLeader
      security:
        - bearerAuth: []
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferClubLeaderRequest'
      responses:
        '200':
          description: 변경 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/ClubDetail'
                  error:
                    nullable: true
        '400':
          description: 자기 자신에게 넘기려 함 (INVALID_LEADER_TRANSFER)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증되지 않음
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
        '403':
          description: 호출한 사용자가 동아리장이 아님 (NOT_CLUB_LEADER)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
        '404':
          description: 존재하지 않는 동아리(CLUB_NOT_FOUND) 또는 대상이 이 동아리의 멤버가 아님(CLUB_MEMBERSHIP_NOT_FOUND)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
  /clubs/{clubId}/applications:
    get:
      summary: 가입 신청 목록 (관리자 전용)
      description: >-
        현재 동아리장만 호출할 수 있다. 아직 처리하지 않은 PENDING 신청서만
        내려준다 — APPROVED는 이미 멤버 목록에서 보이고, REJECTED는 더 이상
        조치할 게 없어서 관리 화면에 남겨둘 필요가 없다.
      operationId: getPendingClubApplications
      security:
        - bearerAuth: []
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ClubApplication'
                  error:
                    nullable: true
        '401':
          description: 인증되지 않음
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
        '403':
          description: 호출한 사용자가 동아리장이 아님 (NOT_CLUB_LEADER)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
        '404':
          description: 존재하지 않는 동아리(CLUB_NOT_FOUND)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
  /clubs/{clubId}/applications/{applicationId}/approve:
    post:
      summary: 가입 신청 승인 (관리자 전용)
      description: >-
        현재 동아리장만 호출할 수 있다. 승인하면 club_members에 행이 생겨
        실제로 가입 처리되고(memberCount +1), 신청서 상태가 APPROVED로 바뀐다.
        이미 APPROVED/REJECTED로 처리된 신청서를 다시 승인하려 하면
        CLUB_APPLICATION_NOT_PENDING(409)을 반환한다.
      operationId: approveClubApplication
      security:
        - bearerAuth: []
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: applicationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 승인 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/ClubApplication'
                  error:
                    nullable: true
        '401':
          description: 인증되지 않음
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
        '403':
          description: 호출한 사용자가 동아리장이 아님 (NOT_CLUB_LEADER)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
        '404':
          description: 존재하지 않는 동아리(CLUB_NOT_FOUND) 또는 신청서(CLUB_APPLICATION_NOT_FOUND)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
        '409':
          description: 이미 처리된 신청서 (CLUB_APPLICATION_NOT_PENDING)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
  /clubs/{clubId}/applications/{applicationId}/reject:
    post:
      summary: 가입 신청 거절 (관리자 전용)
      description: >-
        현재 동아리장만 호출할 수 있다. 거절하면 신청서 상태만 REJECTED로
        바뀌고 club_members에는 아무 변화가 없다(원래 없었으니 그대로).
        지원자는 이후 동아리 상세 화면에서 REJECTED 상태를 보고 재신청할 수
        있다(POST /clubs/{clubId}/apply 참고). 이미 처리된 신청서를 다시
        거절하려 하면 CLUB_APPLICATION_NOT_PENDING(409)을 반환한다.
      operationId: rejectClubApplication
      security:
        - bearerAuth: []
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: applicationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 거절 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/ClubApplication'
                  error:
                    nullable: true
        '401':
          description: 인증되지 않음
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
        '403':
          description: 호출한 사용자가 동아리장이 아님 (NOT_CLUB_LEADER)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
        '404':
          description: 존재하지 않는 동아리(CLUB_NOT_FOUND) 또는 신청서(CLUB_APPLICATION_NOT_FOUND)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
        '409':
          description: 이미 처리된 신청서 (CLUB_APPLICATION_NOT_PENDING)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
...